ip-terminal-code/requirements.txt
RPi.GPIO
RPLCD
smbus2
